Use interface for configuring custom partition key
